Move selectIsLoggedIn selector to auth and adjust tests
