+++ src/containers/PrivateRoute.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { Route, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 
-import { selectGetCurrentUserIsLoading } from 'store/auth/reducer';
-import { selectIsLoggedIn, selectLoggedInUser } from 'store/entities/selectors';
+import { selectGetCurrentUserIsLoading, selectIsLoggedIn } from 'store/auth/reducer';
+import { selectLoggedInUser } from 'store/entities/selectors';
 import LoadingPlaceholder from 'shared/LoadingPlaceholder';
 import getRoleTypesFromRoles from 'utils/user';
 import { UserRolesShape } from 'types';
