+++ src/shared/User/LoginButton.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { get } from 'lodash';
 
-import { selectIsLoggedIn } from 'store/entities/selectors';
+import { selectIsLoggedIn } from 'store/auth/reducer';
 import { isDevelopment } from 'shared/constants';
 import { LogoutUser } from 'utils/api';
 import { logOut } from 'store/auth/actions';
