+++ src/store/entities/selectors.js
@@ -7,10 +7,6 @@ import { createSelector } from 'reselect';
  */
 
 /** User */
-export const selectIsLoggedIn = (state) => {
-  return state.user?.userInfo?.isLoggedIn || null;
-};
-
 export const selectLoggedInUser = (state) => {
   if (state.entities.user) return Object.values(state.entities.user)[0];
   return null;
